Version 1.2 Is LIVE!
Changes:

Removed "Scale X" and "Scale Y" have been removed and replaced with the variable "Shape".  Shape is a Shape2D for the hitbox's collision.  This allows the user to have different types of hitboxes based on the user's preference.

NEW VARIABLE “hit_stun”.  This variable returns a float that can be used for the hit stun of the hitbox.  This variable is very similar to the "Timer" object.

“Hitbox” is now a class type.